requirements.txt
asgiref==3.6.0
Django==3.2.16
django-bootstrap-modal-forms==2.2.0
Pillow==9.4.0
python-decouple==3.7
pytz==2022.7
sqlparse==0.4.3
djangorestframework
markdown
django-filter
drf-yasg
black
djangorestframework-simplejwt
